The hierarchical cache never hit, and the diagnostic cost three passes
Chasing why end-to-end lagged compute-only found two host-side faults, one
of them also unsound.

_gpu_hier cached the coarse templates under a key containing id(H). H is
`self._gtmpl[t0:t0 + nt]`, a fresh numpy view built on every call, so the
key was a new number every time and the cache hit exactly never: it
rebuilt nt x band complex twice per run. At 65536 pairs that was 3.9 ms of
numpy against 2.1 ms of GPU -- the host was doing nearly twice the work of
the device it was feeding.

id() was also the wrong thing to ask. CPython reuses the address of a
freed object, so the next call's view can land on the previous one's id
and hit the cache for a DIFFERENT template sub-range of the same shape,
serving coarse templates built from other templates. H.ctypes.data cannot
collide that way -- it is the address of the templates themselves, so a
different t0 is a different key -- and it is stable across calls, which is
what makes the cache work at all.

The refine-rate bookkeeping then ran three reductions per call: a bool
array, mean(), and sum(). mean is sum/size, so two were free to drop. It
is a diagnostic, and it was costing about 0.1 ms against 0.31 ms of GPU.

Also dropped an astype(int64) in both backends' readback. PEAK_DTYPE's
index field is int64 and assigning int32 into it widens during the strided
write that happens anyway, so the conversion was a whole extra pass and
allocation. Honest note: measured on its own this changed nothing -- the
cast was 0.007 ms. It is a simplification, not the fix.

hier(end), and the end-to-end speedup over one AVX-512 core:

        before          after
  1024  5.909 ms  2.4x  0.501 ms  25.3x
  2048  0.382 ms 17.3x  0.363 ms  18.8x
  4096  0.836 ms  8.2x  0.349 ms  18.7x
  8192  0.315 ms 19.3x  0.254 ms  25.9x
 16384  2.864 ms  5.4x  1.046 ms  14.4x

Host overhead is now about 2.2 ns per pair: 5% of the call at n=16384,
21% at n=4096, and 47% only at n=1024, where the kernel itself is 0.3 ms
and there is very little for it to hide behind.

+++ python/matchedfilter/_mtlcompute.py
@@ -380,7 +380,7 @@ class Context:
         self._check_completed(cmd)
 
         out = nd * nt * nbins
-        idx = b_idx.read(np.int32, out).astype(np.int64).reshape(nd, nt, nbins)
+        idx = b_idx.read(np.int32, out).reshape(nd, nt, nbins)
         val = b_val.read(np.float32, out * 2).view(
             np.complex64).reshape(nd, nt, nbins)
         return idx, val
@@ -520,8 +520,7 @@ class Context:
         self._check_completed(cmd)
 
         out = nd * nt * nbins
-        idx = bufs["idx"].read(np.int32, out).astype(np.int64).reshape(
-            nd, nt, nbins)
+        idx = bufs["idx"].read(np.int32, out).reshape(nd, nt, nbins)
         val = bufs["val"].read(np.float32, out * 2).view(
             np.complex64).reshape(nd, nt, nbins)
         return idx, val
